refactor(syscall): split filesystem syscalls
